requirements.txt
streamlit
tensorflow
pillow
numpy
firebase-admin
flask
flask-cors
opencv-python-headless